Overview

TODO: add example figure here.

If you are new to this method of surrogate time series, feel free to read any of the following:

The method of surrogate testing

What is a surrogate time series?

Types of surrogate realizations

API

TimeseriesSurrogates.jl exports two main functions. Both of them dispatch on the chosen method, a subtype of Surrogate.

TimeseriesSurrogates.surrogate — Function

surrogate(x, method::Surrogate) → s

Create a single surrogate timeseries s from x based on the given method. If you want to generate more than one surrogates from x, you should use surrogenerator.

TimeseriesSurrogates.surrogenerator — Function

surrogenerator(x, method::Surrogate) → sg::SurrogateGenerator

Initialize a generator that creates surrogates of x on demand, based on given method. This is efficient, because for most methods some things can be initialized and reused for every surrogate.

To generate a surrogate, call sg as a function with no arguments, e.g.:

sg = surrogenerator(x, method)
for i in 1:1000
    s = sg()
    # do stuff with s and or x
    result[i] = stuff
end

Surrogate methods

TimeseriesSurrogates.RandomShuffle — Type

RandomShuffle() <: Surrogate

A random constrained surrogate, generated by shifting values around.

This method destroys any linear correlation in the signal, but preserves its amplitude distribution.

TimeseriesSurrogates.BlockShuffle — Type

BlockShuffle(n::Int) <: Surrogate

A block shuffle surrogate constructed by dividing the time series into n blocks of roughly equal width at random indices (end blocks are wrapped around to the start of the time series).

These surrogates preserve short-term correlations in the time series, but break any long-term dynamical information.

TimeseriesSurrogates.RandomFourier — Type

RandomFourier(phases = true) <: Surrogate

A surrogate[Theiler1992] that randomizes the Fourier components of the signal in some manner. If phases==true, the phases are randomized, otherwise the amplitudes are randomized.

If phases==true, then the resulting signal has same linear correlation, or periodogram, as the original data.

#TODO: Okay, what happens if phases!=true ?

TimeseriesSurrogates.AAFT — Type

AAFT()

An amplitude-adjusted-fourier-transform surrogate[Theiler1992].

AAFT have the same linear correlation, or periodogram, and also preserves the amplitude distribution of the original data.

TimeseriesSurrogates.IAAFT — Type

IAAFT(M = 100, tol = 1e-6, W = 75)

An iteratively adjusted amplitude-adjusted-fourier-transform surrogate[SchreiberSchmitz1996].

IAAFT surrogate have the same linear correlation, or periodogram, and also preserves the amplitude distribution of the original data, but are improved relative to AAFT through iterative adjustment (which runs for a maximum of M steps). During the iterative adjustment, the periodograms of the original signal and the surrogate are coarse-grained and the powers are averaged over W equal-width frequency bins. The iteration procedure ends when the relative deviation between the periodograms is less than tol (or when M is reached).

References

TimeseriesSurrogates.PseudoPeriodic — Type

PseudoPeriodic(d, τ, ρ, shift=true) <: Surrogate

Create surrogates suitable for pseudo-periodic signals. They retain the periodic structure of the signal, while inter-cycle dynamics that are either deterministic or correlated noise are destroyed (for appropriate ρ choice). Therefore these surrogates are suitable to test the null hypothesis that the signal is periodic with uncorrelated noise[Small2001].

Arguments d, τ, ρ are as in the paper, the embedding dimension, delay time and noise radius. The method works by performing a delay coordinates ambedding via the library DynamicalSystems.jl. See its documentation for choosing appropriate values for d, τ. For ρ, we have implemented the method proposed in the paper in the function noiseradius.

The argument shift is not discussed in the paper, but it is possible to adjust the algorithm so that there is no phase shift between the periodic component of the original and surrogate data.

Utils

TimeseriesSurrogates.noiseradius — Function

noiseradius(x::AbstractVector, d::Int, τ, ρs, n = 1) → ρ

Use the proposed* algorithm of[Small2001] to estimate optimal ρ value for PseudoPeriodic surrogates, where ρs is a vector of possible ρ values.

*The paper is ambiguous about exactly what to calculate. Here we count how many times we have pairs of length-2 that are identical in x and its surrogate, but are not also part of pairs of length-3.

This function directly returns the arg-maximum of the evaluated distribution of these counts versus ρ, use TimeseriesSurrogates._noiseradius with same arguments to get the actual distribution. n means to repeat τhe evaluation n times, which increases accuracy.

Visualization

TimeseriesSurrogates.jl provides the function surroplot(x, s), which comes into scope when using Plots. This function is used in the example applications.

Theiler1992J. Theiler et al., Physica D 58 (1992) 77-94 (1992)

SchreiberSchmitz1996T. Schreiber; A. Schmitz (1996). "Improved Surrogate Data for Nonlinearity Tests". Phys. Rev. Lett. 77 (4)

Small2001Small et al., Surrogate test for pseudoperiodic time series data, Physical Review Letters, 87(18)

Small2001Small et al., Surrogate test for pseudoperiodic time series data, Physical Review Letters, 87(18)