Truncated FT/AAFT surrogates

TimeseriesSurrogates.TFTS — Type

TFTS(fϵ::Real)

A truncated Fourier transform surrogate[Nakamura2006] (TFTS). These surrogates were designed to deal with data with irregular fluctuations superimposed over long term trends, and are generated by leaving a fraction of the lower frequencies untouched when performing the phase shuffling step (as opposed to randomizing all frequencies, like in RandomFourier surrogates).

The truncation parameter fϵ is the ratio of the high frequency domain to the whole frequency domain. For example, if the number of frequencies is 2000 and frequencies between 1500 and 2000 are to be randomized, then fϵ = 500/2000 = 0.25. The appropriate value of fϵ strongly depends on the data and time series length, and must be manually determined[Nakamura2006], for example by comparing periodograms for the time series and the surrogates.

The TFTS surrogate can be used to test the null hypothesis that the signal is a stationary linear system generated the irregular fluctuations part of the signal[Nakamura2006].

The by changing the sign of fϵ, the implementation here also allows for truncating at the opposite end of the spectrum. Thus, negative values of fϵ will leave the lower frequencies intact and shuffle the higher frequencies (the absolute value of fϵ still refers to the fraction of high frequency domain to total frequency domain).

TimeseriesSurrogates.TAAFT — Type

TAAFT(fϵ)

An truncated version of the amplitude-adjusted-fourier-transform surrogate[^Theiler1992, ^Nakamura2006].

The truncation parameter and phase randomization procedure is identical to TFTS, but here an additional step of rescaling back to the original data is performed. This preserves the amplitude distribution of the original data.

References

Nakamura2006Nakamura, Tomomichi, Michael Small, and Yoshito Hirata. "Testing for nonlinearity in irregular fluctuations with long-term trends." Physical Review E 74.2 (2006): 026205.

Theiler1992J. Theiler et al., Physica D 58 (1992) 77-94 (1992)](https://www.sciencedirect.com/science/article/pii/016727899290102S)

Nakamura2006Nakamura, Tomomichi, Michael Small, and Yoshito Hirata. "Testing for nonlinearity in irregular fluctuations with long-term trends." Physical Review E 74.2 (2006): 026205.