Example systems

Example processes

TimeseriesSurrogates.NLNS — Function.

NLNS(n_steps, x₀)

A nonlinear, nonstationary process given by the following stochastic model [1]:

$x_t = x_{t-1} + a_t a_{t-1}$

with $a_t ~ N(0,1)$ being stationary. The nonlinearity lies in the $a_t a_{t-1}$ term. This is, in essence, a nonlinear random walk. x₀ sets the initial condition.

Literature references

Lucio et al., Phys. Rev. E 85, 056202 (2012).

source

TimeseriesSurrogates.SNLST — Function.

SNLST(n_steps, x₀, k)

Dynamically linear process transformed by a strongly nonlinear static transformation (SNLST), given by the following map [1]:

$x(t) = k*x(t-1) + a(t)$

with the transformation $s(t) = x(t)^3$.

Literature references

Lucio et al., Phys. Rev. E 85, 056202 (2012).

source

TimeseriesSurrogates.randomwalk — Function.

randomwalk(n_steps, x₀)

Linear random walk (AR(1) process with a unit root) [1]

This is an example of a nonstationary linear process.

Literature references

Lucio et al., Phys. Rev. E 85, 056202 (2012).

source

TimeseriesSurrogates.NSAR2 — Function.

NSAR2(n_steps, x₀, x₁)

Cyclostationary AR(2) process given by the following map [2]

$x_t = a₁(t) * x_{t-1} + a₋ x_{t-2} + ϵ_t$.

where

$a₁(t) = 2 cos[2pi / T(t)] * exp(-1 / τ)$,

$T(t) = T_0 + M * sin(2 pi t / T_{mod})$,

$a₂ = exp(-2 / au)$,

and x₀ and x₁ sets the initial conditions.

Literature references

Lucio et al., Phys. Rev. E 85, 056202 (2012), after J. Timmer, Phys. Rev. E 58, 5153

(1998).

source

TimeseriesSurrogates.AR1 — Function.

AR1(n_steps, x₀, k)

Simple AR(1) model with no static transformation given by the following map [1]:

$x[i] = k*x[i-1] + a[i]$,

where a[i] is a draw from a normal distribution with zero mean and unit variance. x₀ sets the initial condition and k is the tunable parameter in the map.

Literature references

Lucio et al., Phys. Rev. E 85, 056202 (2012).

source