Amplitude adjusted Fourier (AAFT)

Amplitude adjusted Fourier transform surrogates

Different variants of the AAFT and iterated AAFT algorithms.

Amplitude adjusted Fourier transform (AAFT)

TimeseriesSurrogates.aaft — Function.

aaft(ts)

Generate a realization of an amplitude adjusted Fourier transform (AAFT) surrogate.

ts Is the time series for which to generate an AAFT surrogate realization.

Literature references

J. Theiler et al., Physica D 58 (1992) 77-94 (1992).

source

Example of AAFT

using TimeseriesSurrogates
ts = AR1() # create a realization of a random AR(1) process
surrogate = aaft(ts)

surrplot(ts, surrogate)

Iterated AAFT (AAFT)

The IAAFT surrogates add an iterative step to the AAFT algorithm improve convergence.

TimeseriesSurrogates.iaaft — Function.

Generate an iteratively adjusted amplitude adjusted Fourier transform (IAAFT) surrogate series for ts.

From T. Schreiber; A. Schmitz (1996). "Improved Surrogate Data for Nonlinearity Tests". Phys. Rev. Lett. 77 (4): 635–638. doi:10.1103/PhysRevLett.77.635. PMID

source

Example of IAAFT

using TimeseriesSurrogates
ts = AR1() # create a realization of a random AR(1) process
surrogate = iaaft(ts)

surrplot(ts, surrogate)