Random shuffle (RS)

Random shuffle surrogates (RS)

The easiest way of constructing a constrained surrogate is just shuffling the time indices of the original time series.

TimeseriesSurrogates.randomshuffle — Function.

randomshuffle(ts)

Generate a random constrained surrogate for ts.

source

Example of surrogate generation by shuffling time indices

using TimeseriesSurrogates
ts = AR1()
surrogate = randomshuffle(ts)

surrplot(ts, surrogate) # Visualize the surrogate